create AppShell component
