+++ layouts/appShell/index.ts
@@ -1,0 +1,2 @@
+import AppShell from "./AppShell";
+export default AppShell;
