--- pages/_app.css.ts
@@ -1,17 +1,0 @@
-import styled from "styled-components";
-
-export default styled.div`
-  > footer {
-    position: relative;
-    bottom: 30px;
-    margin: 0 auto;
-    text-align: center;
-    left: 0px;
-    right: 0px;
-    padding: 48px;
-  }
-
-  > footer p {
-    color: var(--gray-500);
-  }
-`;
